Feature kitosudv 242 cover app layer code

+++ Core.ApplicationServices/Interface/ExhibitUsage/InterfaceExhibitUsageService.cs
@@ -1,4 +1,5 @@
-﻿using Core.ApplicationServices.Model.Result;
+﻿using Core.ApplicationServices.Authorization;
+using Core.ApplicationServices.Model.Result;
 using Core.DomainModel.ItSystemUsage;
 using Core.DomainServices;
 using Serilog;
@@ -9,13 +10,16 @@ namespace Core.ApplicationServices.Interface.ExhibitUsage
     {
         private readonly IGenericRepository<ItInterfaceExhibitUsage> _itInterfaceExhibitUsageRepository;
         private readonly ILogger _logger;
+        private readonly IAuthorizationContext _authorizationContext;
 
         public InterfaceExhibitUsageService(
             IGenericRepository<ItInterfaceExhibitUsage> itInterfaceExhibitUsageRepository,
-            ILogger logger)
+            ILogger logger,
+            IAuthorizationContext authorizationContext)
         {
             _itInterfaceExhibitUsageRepository = itInterfaceExhibitUsageRepository;
             _logger = logger;
+            _authorizationContext = authorizationContext;
         }
 
         public OperationResult Delete(int systemUsageId, int interfaceExhibitId)
@@ -27,9 +31,22 @@ namespace Core.ApplicationServices.Interface.ExhibitUsage
                 _logger.Error($"Could not find interface exhibit usage with key {key}");
                 return OperationResult.NotFound;
             }
+
+            if (!AllowDelete(interfaceExhibitUsageToBeDeleted))
+            {
+                return OperationResult.Forbidden;
+            }
+
             _itInterfaceExhibitUsageRepository.Delete(interfaceExhibitUsageToBeDeleted);
             _itInterfaceExhibitUsageRepository.Save();
             return OperationResult.Ok;
+        }
+
+        private bool AllowDelete(ItInterfaceExhibitUsage interfaceExhibitUsageToBeDeleted)
+        {
+            //ExhibitUsage belongs to a contract, hence a deletion is a modification of the contract
+            return interfaceExhibitUsageToBeDeleted.ItContract == null || 
+                   _authorizationContext.AllowModify(interfaceExhibitUsageToBeDeleted.ItContract);
         }
     }
 }
